+++ src/decoder.c
@@ -4,7 +4,7 @@
 #define BLACK 0
 #define MAX_ABS_OFFSET 2
 
-int decoder_init(_decoder *self, int iw, int ih, float rw, float rh, 
+int decoder_init(_decoder *self, int iw, int ih, double rw, double rh, 
         _complex *c, _complex *r) {
 	if (self == 0 || c == 0 || r == 0)  //Cant be nulls
 		return -1;
@@ -24,10 +24,10 @@ int decoder_init(_decoder *self, int iw, int ih, float rw, float rh,
 
 int decoder_decode(_decoder *self, FILE *output) {
 	//Concurrency is for poor computers
-	float startX, startY, endX, endY, stepX, stepY;
+	double startX, startY, endX, endY, stepX, stepY;
 
-	float halfWidth = self->renderWidth / 2;
-	float halfHeight = self->renderHeight /2;
+	double halfWidth = self->renderWidth / 2;
+	double halfHeight = self->renderHeight /2;
 	startX = complex_getX(self->renderCenter) - (halfWidth);
 	startY = complex_getY(self->renderCenter) - (halfHeight);
 	endX = complex_getX(self->renderCenter) + (halfWidth);
@@ -40,8 +40,8 @@ int decoder_decode(_decoder *self, FILE *output) {
 	fprintf(output, "%d %d ", self->imageWidth, self->imageHeight);
 	fprintf(output, "%d\n", WHITE);
 
-	for (float indexY = startY ; indexY < endY ; indexY += stepY) {
-		for (float indexX = startX ; indexX < endX ; indexX += stepX) {
+	for (double indexY = startY ; indexY < endY ; indexY += stepY) {
+		for (double indexX = startX ; indexX < endX ; indexX += stepX) {
 			//Here im at 1 px of the image.
 			_complex point;
 			complex_init(&point, indexX, indexY);
@@ -51,10 +51,10 @@ int decoder_decode(_decoder *self, FILE *output) {
 			for (color = BLACK ; 
 				color < WHITE - 1 && complex_abs(&point) <= MAX_ABS_OFFSET ;
 					++color) {
-				float newX = (complex_getX(&point) * complex_getX(&point)
+				double newX = (complex_getX(&point) * complex_getX(&point)
 						- complex_getY(&point) * complex_getY(&point)
 						+ complex_getX(self->ratio));
-				float newY = (2 * complex_getX(&point) * complex_getY(&point)
+				double newY = (2 * complex_getX(&point) * complex_getY(&point)
 						+ complex_getY(self->ratio));
 				
 				complex_init(&point, newX, newY);
